+++ RELEASE_NOTES.md
@@ -3,6 +3,11 @@
 Release notes for the AD Build System (Software Factory)
 
 ## Releases:
+* tag:1.0.0 24-Mar-2025 P. Nisperos (pnispero)
+  * Added deployment type to a component's config.yaml. Updated --new to --update-db for deployment cli.
+  * Added sanity check for deployment to ALL iocs - ensure database is the same as component source tree.
+  * Point CLI and deployment controller to production cluster now.
+
 * tag:1.0.6b0 13-Mar-2025 P. Nisperos (pnispero)
   * Updated deployment controller with bug fixes with mount paths, ansible ssh pipelining. And more error checking in deployment cli.